% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_data_dictionary.R
\name{print_data_dictionary}
\alias{print_data_dictionary}
\title{Print data dictionary as a nicely formatted \code{gt} table.}
\usage{
print_data_dictionary(data)
}
\arguments{
\item{data}{Data.frame. Dataset that has attributes added, including a data dictionary.}
}
\value{
This function will return a \code{gt} object.
}
\description{
\code{print_data_dictionary} takes a data frame to which a data dictionary has
been added using the \code{gt} package. For the function to run, the following parameters are needed.
}
\examples{
# example original data set for which a dictionary will be made
data("esoph")
my.data <- esoph

my.data <- add_data_dictionary(
  data = my.data,
  main_string = "This dataset describes tobacco and alcohol consumption at different age groups.",
  variable_description = c(
    "age group in years", 
    "alcohol consumption in gm/day", 
    "tobacco consumption in gm/day", 
    "number of cases (showing a range)", 
    "number of controls (showing range)"
  ),
  variable_type = c(0, 0, 0, 0, 0),
  na.rm = FALSE
)

print_data_dictionary(my.data)

}
