% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prompt_attr.R
\name{prompt_attr}
\alias{prompt_attr}
\title{Incorporate attributes as metadata to an original dataset as prompted by the 
function.}
\usage{
prompt_attr(my.data, data.dictionary)
}
\arguments{
\item{my.data}{Data.frame. The data set to add attributes as metadata.}

\item{data.dictionary}{Data frame. The data dictionary has all variable names, 
and variable descriptions that will explain an original dataset.}
}
\value{
This function will return an R dataset containing metadata stored in its 
attributes. The function will prompt the user for a main description. Attributes 
added will include: a data dictionary, the name of the author or user who 
created the dictionary and added it, the time when it was last edited and a
brief description of the original dataset.
}
\description{
\code{prompt_attr} adds attributes to an original dataset as metadata, 
including a data dictionary, among other attributes as prompted by the function.
}
\details{
This is a variation of the third function used in this package. For the function 
to run, the following parameters are needed.
}
